Move optimize bools pass to a new file
No code changes, just moving.
