+++ src/coreclr/jit/CMakeLists.txt
@@ -151,6 +151,7 @@ set( JIT_SOURCES
   morphblock.cpp
   objectalloc.cpp
   optcse.cpp
+  optimizebools.cpp
   optimizer.cpp
   patchpoint.cpp
   phase.cpp
